% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_plot.R
\name{amino_acid_plot}
\alias{amino_acid_plot}
\title{Amino Acid Frequency PLot}
\usage{
amino_acid_plot(peptide_sequence)
}
\arguments{
\item{peptide_sequence}{The peptide sequence to be visualized}
}
\value{
The frequency plot for the different amino acids present
}
\description{
Amino Acid Frequency PLot
}
