% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{Function 4: Assemble sequences}
\usage{
assemble(codons)
}
\arguments{
\item{codons}{}
}
\value{
codon values
}
\description{
Function 4: Assemble sequences
}
\examples{
assemble("CCG")
}
