% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codonTable}
\alias{codonTable}
\title{Title}
\format{
here is our table
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codonTable
}
\description{
Description
}
\keyword{datasets}
