% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_seq.R
\name{dna_seq}
\alias{dna_seq}
\title{Generating random DNA sequence}
\usage{
dna_seq(seq_length)
}
\arguments{
\item{seq_length}{The length of the DNA sequence that we want to generate}
}
\value{
A random DNA sequence of length "seq_length"
}
\description{
Generating random DNA sequence
}
\examples{
dna_seq(10)
dna_seq(5)
dna_seq(20)
}
