% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_codons.R
\name{DNA_to_codons}
\alias{DNA_to_codons}
\title{From DNA to codons
Divides the DNA sequence into groups of three}
\usage{
DNA_to_codons(DNA_seq, start = 1)
}
\arguments{
\item{DNA_seq}{sting of DNA sequence}

\item{Length_DNA}{the number of nucleotides in the DNA sequence}
}
\value{
codons a sting of 3 grouped nucleotides
}
\description{
From DNA to codons
Divides the DNA sequence into groups of three
}
