% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{RNA sequence to DNA sequence}
\usage{
dna_to_rna(dna_seq)
}
\arguments{
\item{dna_seq}{}
}
\value{
The corresponding RNA sequence obtained from the DNA sequence
}
\description{
RNA sequence to DNA sequence
}
\examples{
dna_to_rna("ATCGGGGATC")
}
