% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodon.R
\name{getCodon}
\alias{getCodon}
\title{Getting Codins}
\usage{
getCodon(RNASeq, start = 1)
}
\arguments{
\item{start}{Position to start the codon counting/dividing. Default start = 1}

\item{RNAseq}{RNA sequence input}
}
\value{
Goes through a RNA sequence and create the triplet codons (extracting character vector substring by 3)
}
\description{
Getting Codins
}
