% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dna.R
\name{make_dna}
\alias{make_dna}
\title{Create DNA string}
\usage{
make_dna(dna_length)
}
\arguments{
\item{dna_length}{choose the length of the wanted DNA string}
}
\value{
The function gives a random sampled DNA string with a given length
}
\description{
Create DNA string
}
\examples{
make_dna(8)
}
