% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateRNA.R
\name{protein_seq}
\alias{protein_seq}
\title{Create protein sequence based on RNA sequence}
\usage{
protein_seq(input_codons)
}
\arguments{
\item{The}{parameters should be RNA codons}
}
\value{
Protein sequence based on RNA sequence
}
\description{
Create protein sequence based on RNA sequence
}
\examples{
test_rna <- c("AUG", "AUG")
protein_seq(test_rna)

}
