% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_subs.R
\name{RNA_subs}
\alias{RNA_subs}
\title{Substitute Thymin for Uracil}
\usage{
RNA_subs(DNA_data)
}
\arguments{
\item{DNA_data}{data set of DNA nucleotides}
}
\value{
returns a data set of RNA nucleotides
}
\description{
Substitute Thymin for Uracil
}
\examples{
data <- c("A", "T", "T", "C")
RNA_subs(data)

}
