% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_to_codon.R
\name{sequence_to_codon}
\alias{sequence_to_codon}
\title{Title sequence_to_codons}
\usage{
sequence_to_codon(main_sequence, start = 1)
}
\arguments{
\item{main_sequence}{sequence to be converted into codons}

\item{start}{}
}
\value{
codons i.e, multiple sequences of three nucleotides from the main sequence. It returns a substring
}
\description{
Title sequence_to_codons
}
\examples{
main_sequence = "ATCGAT"
}
