% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringDNA.R
\name{stringDNA}
\alias{stringDNA}
\title{stringDNA}
\usage{
stringDNA(size = 1)
}
\arguments{
\item{size}{Integer defining length of DNA string build of A, T, G and C. (Default is 1)}
}
\value{
A string of DNA characters.
}
\description{
Function that generates a string of DNA bases with size given as function argument.
}
\examples{
stringDNA(size = 20)
}
