% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Transcribe DNA sequence to RNA sequence}
\usage{
transcribe(dna_seq)
}
\arguments{
\item{dna_seq}{The DNA sequence to be converted}
}
\value{
The corresponding RNA sequence obtained from the DNA sequence
}
\description{
Transcribe DNA sequence to RNA sequence
}
\examples{
transcribe("ATCGGGGATC")
}
