% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Function 4: Assemble sequences}
\usage{
translate(codons)
}
\arguments{
\item{codons}{}
}
\value{
codon values
}
\description{
Function 4: Assemble sequences
}
\examples{
assemble("CCG")
}
