% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodon.R
\name{getCodon}
\alias{getCodon}
\title{Getting Codons}
\usage{
getCodon(RNASeq, start = 1)
}
\arguments{
\item{start}{Position to start the codon counting/dividing. Default start = 1}

\item{RNAseq}{RNA sequence input as string.}
}
\value{
Character vector where each element is a codon.
}
\description{
Getting Codons
}
\examples{
getCodon("UCGUUA", start = 1). Returns a character vector c("CGU", "UAA")
}
