% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Function 4: Translate sequences}
\usage{
translate(codons)
}
\arguments{
\item{codons}{sequence of codons}
}
\value{
corresponding amino acid code
}
\description{
Function 4: Translate sequences
}
\examples{
translate("CCG")
}
